Added Ambiq SDK
 * copied from what is downloaded as part of a TFLM build
 * removed the unnecessary bits to reduce the size from 562M to 7.4M

The idea here is that all third_party libraries that are not needed for
code TFLM (but needed to run examples on a specific target) should be
directly added to the target specific examples repository.

It is up to the maintainer of the target-specific examples repository
how they want to handle these third_party libraries (i.e. download as
part of the build, check into the repo directly ...)